+++ __tests__app__/app/entry.js
@@ -12,4 +12,6 @@ Socket.setMaxMessageSize((2 ** 20) * 100);
 console.log('Custom entry', 'imported');
 Application.addListener('start', () => {
 	console.log('Custom entry', 'Application started.');
+
+	// Application.setLocale('default');
 });